+++ src/constant.h
@@ -218,8 +218,6 @@ constexpr const char *LOCAL_CONFIG_DIR = "~/.ydsh";
 constexpr const char *LOCAL_MOD_DIR = "~/.ydsh/module";
 
 constexpr const char *SYSTEM_DATA_DIR = X_DATADIR "/ydsh";
-
-constexpr const char *SYSTEM_CONFIG_DIR = X_SYSCONFIGDIR "/ydsh";
 constexpr const char *SYSTEM_MOD_DIR = X_DATADIR "/ydsh/module";
 
 } // namespace ydsh